requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyJWT
cryptography
google-genai
